Change abtract factory for pimpl
The original idea was to abtract GLFW dependencies so it can be changed for something else at runtime, (Other window/keyboard/mouse/... implementations) but there will really be only one implementation anyway. So I removed all the complexity of (probably useless) flexibility for a simple pimpl and a good interface. This way I may change the implementation (e.g I don't like GLFW anymore and want to use SDL2) later but without touching the interface so the client is happy.